Create loader information for school page
