+++ web/src/pages/Home.jsx
@@ -1,59 +1,14 @@
 import Searchbar from "../components/Searchbar"
 import QA from "../components/QA"
 import { useEffect, useState } from "react"
-import api from "../utilities/Axios"
+import { getSchools } from "../utilities/GetData"
 
 const Home = () => {
   const [schools, setSchools] = useState([])
-  const [courses, setCourses] = useState([])
 
   const fetchSchools = async () => {
-    try {
-      const response = await api.get("/school/")
-      if (response && response.data) {
-        const resultingArray = response.data.data.map((curr) => ({
-          long_name: curr.long_name,
-          short_name: curr.short_name,
-          value: curr.short_name,
-          label: curr.long_name,
-        }))
-
-        setSchools(resultingArray)
-      }
-    } catch (err) {
-      if (err.response) {
-        // Not in the 200 response range
-        console.log(err.response.data)
-        console.log(err.response.status)
-        console.log(err.response.headers)
-      } else {
-        console.log(`Error: ${err.message}`)
-      }
-    }
-  }
-
-  const fetchCourses = async () => {
-    // try {
-    //   const response = await api.get("/schools/")
-    //   if (response && response.data) {
-    //     const resultingArray = response.data.data.map((curr) => ({
-    //       long_name: curr.long_name,
-    //       short_name: curr.short_name,
-    //       value: curr.short_name,
-    //       label: curr.long_name,
-    //     }))
-    //     setSchools(resultingArray)
-    //   }
-    // } catch (err) {
-    //   if (err.response) {
-    //     // Not in the 200 response range
-    //     console.log(err.response.data)
-    //     console.log(err.response.status)
-    //     console.log(err.response.headers)
-    //   } else {
-    //     console.log(`Error: ${err.message}`)
-    //   }
-    // }
+    const response = await getSchools()
+    setSchools(response)
   }
 
   useEffect(() => {
